+++ src/test/java/org/tinyradius/core/attribute/type/IntegerAttributeTest.java
@@ -12,7 +12,7 @@ class IntegerAttributeTest {
 
     @Test
     void intMaxUnsigned() {
-        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, (byte) 27, -1);
+        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, 27, -1);
 
         assertEquals(-1, attribute.getValueInt());
         assertEquals(0xffffffffL, attribute.getValueLong());
@@ -22,7 +22,7 @@ class IntegerAttributeTest {
     @Test
     void intMaxSigned() {
         final int value = Integer.MAX_VALUE;
-        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, (byte) 27, value);
+        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, 27, value);
 
         assertEquals(value, attribute.getValueInt());
         assertEquals(value, attribute.getValueLong());
@@ -32,7 +32,7 @@ class IntegerAttributeTest {
     @Test
     void bytesOk() {
         final byte[] bytes = new byte[]{0, 0, (byte) 0xff, (byte) 0xff};
-        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, (byte) 10, bytes);
+        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, 10, bytes);
         assertEquals(65535, attribute.getValueInt());
         assertEquals(65535, attribute.getValueLong());
         assertEquals("65535", attribute.getValueString());
@@ -41,20 +41,20 @@ class IntegerAttributeTest {
     @Test
     void bytesTooShort() {
         final IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> new IntegerAttribute(dictionary, -1, (byte) 10, new byte[2]));
+                () -> new IntegerAttribute(dictionary, -1, 10, new byte[2]));
         assertTrue(exception.getMessage().toLowerCase().contains("should be 4 octets"));
     }
 
     @Test
     void bytesTooLong() {
         final IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> new IntegerAttribute(dictionary, -1, (byte) 10, new byte[5]));
+                () -> new IntegerAttribute(dictionary, -1, 10, new byte[5]));
         assertTrue(exception.getMessage().toLowerCase().contains("should be 4 octets"));
     }
 
     @Test
     void stringOk() {
-        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, (byte) 10, "12345");
+        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, 10, "12345");
         assertEquals(12345, attribute.getValueInt());
         assertEquals(12345, attribute.getValueLong());
         assertEquals("12345", attribute.getValueString());
@@ -64,7 +64,7 @@ class IntegerAttributeTest {
     void stringTooBig() {
         final String strLong = Long.toString(0xffffffffffL);
         final NumberFormatException exception = assertThrows(NumberFormatException.class,
-                () -> new IntegerAttribute(dictionary, -1, (byte) 27, strLong));
+                () -> new IntegerAttribute(dictionary, -1, 27, strLong));
 
         assertTrue(exception.getMessage().toLowerCase().contains("exceeds range"));
     }
@@ -72,14 +72,14 @@ class IntegerAttributeTest {
     @Test
     void stringEmpty() {
         final NumberFormatException exception = assertThrows(NumberFormatException.class,
-                () -> new IntegerAttribute(dictionary, -1, (byte) 27, ""));
+                () -> new IntegerAttribute(dictionary, -1, 27, ""));
 
         assertTrue(exception.getMessage().toLowerCase().contains("for input string: \"\""));
     }
 
     @Test
     void stringEnum() {
-        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, (byte) 6, "Login-User");
+        final IntegerAttribute attribute = new IntegerAttribute(dictionary, -1, 6, "Login-User");
 
         assertEquals(1, attribute.getValueInt());
         assertEquals(1, attribute.getValueLong());
@@ -89,7 +89,7 @@ class IntegerAttributeTest {
     @Test
     void stringInvalid() {
         final NumberFormatException exception = assertThrows(NumberFormatException.class,
-                () -> new IntegerAttribute(dictionary, -1, (byte) 6, "badString"));
+                () -> new IntegerAttribute(dictionary, -1, 6, "badString"));
 
         assertTrue(exception.getMessage().toLowerCase().contains("for input string: \"badstring\""));
     }
